Adding ticket locking logic to be handled
Tickets created/updated from must be locked if the order is still not
in its confirmed state. It will not allow other users to operate on the
same ticket which is currently being processed by another user. This is
achieved by maintaining maintaining orderID field in the tickets. If
orderId in ticket document is not set, it means the ticket is not locked
currently. Creating listeners in tickets to get events from Orders
service that will update this orderId accordingly.